Remove library version specification
When I tried to install the package on the HPC I use I didn't want to update packages to newer versions incase it had unforseen consequences. I have run an earlier version of this package on the HPC and therefore the current versions of libraries must be fine. I don't know how early a version would be good enough.